:recycle: 이종수: index 조회 기능 리팩토링 완료
